add typing indicator component
